chore: extract tree-related logic to a new library
